[PATCH] WifiStateMachine: Probe neighbours only when roaming
Whenever a WPA group rekeying happens, the wpa_supplicant state goes
from CONNETED to GROUP_HANDSHAKE and back to CONNETED. This makes
WifiStateMachine call IpManager.confirmConfiguration() which, in this
scenario, leads to the NUD_FAILED state and subsequent disconnection.
Instead of calling IpManager.confirmConfiguration() whenever we reach
the CONNECTED state, do it only if we are roaming.